+++ src/main/java/org/thema/pixscape/view/ComputeViewJava.java
@@ -107,7 +107,7 @@ public final class ComputeViewJava extends SimpleComputeView {
      * @param startZ the height of the eye
      * @param bounds the limits of the view
      * @param view the resulting tangential view
-     * @param a the horizontal angle of the ray
+     * @param a the horizontal angle of the ray in (0-2PI(
      * @param wa the width of the view buffer
      * @param ax the x index in the view buffer for this ray
      * @param ares the resolution (in degree) of the view
@@ -119,16 +119,17 @@ public final class ComputeViewJava extends SimpleComputeView {
         final double res = getData().getResolution();
         int y1 = a >= 0 && a < Math.PI ? 0 : h-1; // haut ou bas ?
         int x1 = a >= Math.PI/2 && a < 1.5*Math.PI ? 0 : w-1; // droite ou gauche ?
-        int sens = x1 == 0 ? -1 : +1;
-
-        int ddy = (int) -Math.round(Math.tan(a) * Math.abs(x1-c0.x));
-        int y = c0.y + sens * ddy;   
+        
+        int ddy = (int) Math.round(Math.tan(a) * (c0.x-x1));
+        int y = c0.y + ddy;   
         if(y >= 0 && y < h) {
             y1 = y;   
         } else {
-            int ddx = (int) Math.abs(Math.round(Math.tan(a+Math.PI/2) * Math.abs(y1-c0.y)));
-            x1 = c0.x + sens * ddx;
-        }        
+            int ddx = (int) Math.round((c0.y-y1) / Math.tan(a));
+            x1 = c0.x + ddx;
+        }
+
+        System.out.println(x1 + " - " + y1 + " - a:" + a/Math.PI);
         
         final double z0 = dtm.getSample(c0.x, c0.y, 0) + startZ;
         
@@ -165,6 +166,9 @@ public final class ComputeViewJava extends SimpleComputeView {
                 ind += sy*w;
             }
 
+            if(ind >= dtmBuf.length) {
+                ind--;
+            }
             double z = dtmBuf[ind] + (dsmBuf != null ? dsmBuf.getElemDouble(ind) : 0);
             if(Double.isNaN(z)) {
                 return;